Add build-only option to deploy

+++ deploy.sh
@@ -1,6 +1,9 @@
 set -eo pipefail
 shopt -s dotglob
+
 rm -r build/* || true
 shopt -u dotglob
 grunt
-rsync -rLkpEtzPc --delete-delay build/ rcchan@keating:~/web/
+if [[ "$1" != "--build" ]]; then
+  rsync -rLkpEtzPc --delete-delay build/ rcchan@keating:~/web/
+fi
